+++ src/main/java/net/basilisk/heartofscales/registry/ModCreativeTabs.java
@@ -30,6 +30,7 @@ public final class ModCreativeTabs {
                         output.accept(ModItems.DRAGON_SPAWN_EGG.get());
                         output.accept(ModItems.DRAGON_BEACON.get());
                         output.accept(ModItems.DRAGON_STAFF.get());
+                        output.accept(ModItems.DRAGON_SADDLE.get());
                     })
                     .build());